requirements.txt
streamlit==1.12.2
google-generativeai==0.1.0
Pillow==9.2.0
pyttsx3==2.90